Revert changes to diffChildren from #159 - Simplify finding oldVNode in diffChildren (+14 B)
It turns out that in [diffchildren-index-tracing-2](https://github.com/developit/ceviche/tree/diffchildren-index-tracing-2)
(which I'd like to open a PR for), I have to bring back the index
tracking that #159 removed. So before I open a PR for
diffchildren-index-tracing-2, I'd like to revert the changes to
diffChildren from #159 so that it is easier to see the real changes that
the index tracing introduces.