Remove stateChanged tracking (-23 B)
Instead of manually forcing an update for layout effects, just queue the state updates to the standard update queue. For standard usages of Preact, this will be flushed in the microTask queue, which is accomplishes the goal of layout effects (that they are invoked before the browser renders). This has the benefit that layout effect updates are batched with other updates, reducing the normal of total rerenders.