Switch from *.mjs -> *.module.js (#1425)
This PR basically reverts #1285 and moves back to `.js` files. It reduces friction with webpack which doesn't resolve `.mjs` the same way as `.js` thus breaking imports in submodules. We've been pointing users to the excellent [webpack-modules](https://github.com/lukeed/webpack-modules) plugin which corrects that as a current workaround.

This PR should remove the need for that plugin making preact work right out of the box with any webpack configuration.

Fixes #1424 , #1321 and lots of questions on slack :+1: