Remove more null checks from diff (-13 B)
The only place `diff` is called not inside `diffChildren` (which already does proper null checking) is `forceUpdate`. `forceUpdate` will never pass in a null newVNode or oldVNode cuz it always passes the VNode that represents the component. If the component is in the tree, it can't have a null VNode (or else it doesn't exist in the tree. It can render null, but that is handled by `diffChildren`).