Add toEqualNode chai assertion (#1680)
Add a new chai assertion to check if two DOM Nodes are equal. This helper improves the error output if the assertion is incorrect by first checking if the tagNames of the two Nodes are equal before checking their referential equality.

Makes it unlikely that Karma will dump the entire DOM contents of a Node if a test is failing.