Delete React debugging props __self and __source
The @babel/preset-react preset includes two plugins which add `__source` and `__self` attributes to every element included development, which React uses for debugging. This muddles up the DOM and is pretty annoying.

For users of preact/compat, it is hard two remove these plugins as Babel does not allow disabling plugins included in a preset (see https://github.com/babel/babel/issues/3016#issuecomment-156150431), and forking `babel-preset-react-app` would mean having to keep updating it with upstream changes.

This is a tiny patch that, in development only, deletes these two attributes from every VNode if they exist. This should not impact production bundle size or performance, but would be a big win for the DX for our project.