Remove throwing suspense test and unused code
It seems  in React a thrown promise that is rejected and not handled in user code just causes an infinite loop (https://codesandbox.io/s/react-failing-lazy-component-h1q22). In other words, by default, all react does on thrown promises is trigger a re-render. It is up to the component that threw the promise (.e.g. lazy or react-cache) to properly handle the rejection. Since user-thrown Promises aren't a supported scenario in React, I don't think we should support them either