Schedule effect/update queue flushes if hooks change
When scheduling a state update or effect, check whether the
`options.{debounceRendering, requestAnimationFrame}` hooks changed since
an update/effect queue flush was last scheduled and schedule a flush
using the new hooks if so.

This fixes an issue where a call to `act` would fail to synchronously flush
state updates and effects if a flush was already scheduled but not yet executed
before `act` was called. This happened because when a state update or effect was
queued during the `act` callback, if the global update/effect queues were
non-empty then the temporary hooks installed by `act` to facilitate
synchronously flushing the queue would not be called.

More generally, this fixes an issue where any changes to the scheduling hooks
do not take effect until any already-scheduled flushes have happened.

Fixes #1794