Fix argumentless useLayoutEffect debug test
The previous test passed because the diffed option was deleted before each test, which deleted the hook.diffed option. In diffed, the hook code flushes useLayoutEffect so in a real app, the debug code would always see any layout effects because it ran after they were run. Because the old test had deleted the hooks option, this flushing code never ran and the debug option always saw the layoutEffects