WIP: Use getDomSibling to update oldDom if a child dom goes null
A test is failing cuz when trying to find the sibling dom in `getDomSibling`, a parent is in the middle of diffing and it's `_children` prop is set to it's render result, which is not an array.

The `_parent` property is shared between oldVNode and newVNode in this situation so when traversing up oldVNode `_parent`, you reach a parent's newVNode. In other words, in `renderComponent` , `/* oldVNode */ assign({}, vnode)._children[0]._parent == /* newVNode */ vnode`